feat: integrate expo-router and enhance app structure with new layout and ripple effect
