+++ example/app.json
@@ -3,6 +3,7 @@
     "name": "example",
     "slug": "example",
     "version": "1.0.0",
+    "scheme": "rnchessboard",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "light",
@@ -28,6 +29,7 @@
     },
     "web": {
       "favicon": "./assets/favicon.png"
-    }
+    },
+    "plugins": ["expo-router"]
   }
 }
