--- example/index.js
@@ -1,5 +1,0 @@
-import { registerRootComponent } from 'expo';
-
-import App from './App';
-
-registerRootComponent(App);
