+++ example/package.json
@@ -1,7 +1,7 @@
 {
   "name": "example",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo run:android",
@@ -11,13 +11,17 @@
   "dependencies": {
     "@shopify/react-native-skia": "2.4.18",
     "expo": "~55.0.0",
+    "expo-constants": "~55.0.16",
     "expo-linking": "^55.0.15",
+    "expo-router": "~55.0.16",
     "expo-status-bar": "~55.0.5",
     "react": "19.2.0",
     "react-dom": "19.2.0",
     "react-native": "0.83.6",
     "react-native-gesture-handler": "~2.30.0",
     "react-native-reanimated": "4.2.1",
+    "react-native-safe-area-context": "~5.6.2",
+    "react-native-screens": "~4.23.0",
     "react-native-web": "^0.21.0",
     "react-native-worklets": "0.7.4"
   },
